Add Triggerise departmental rules

+++ db/migrate/20180428130158_create_items.rb
@@ -3,7 +3,7 @@ class CreateItems < ActiveRecord::Migration[5.1]
     create_table :items do |t|
       t.belongs_to :product, foreign_key: true
       t.belongs_to :sale, foreign_key: true
-      t.decimal :quantity
+      t.integer :quantity
       t.decimal :total
 
       t.timestamps
